Change time_resolution to operate in micro_s, address PR comments
