--- operations_simulator/plumbing/config.py
@@ -1,5 +1,0 @@
-'''config settings'''
-DEFAULT_TIME_RESOLUTION = 0.5
-DEFAULT_RESOLUTION_SCALE = 20
-FC_MAX = 10
-CLOSED_KEYWORD = 'closed'
